Expose the script settings
