--- WinIRC/Views/Settings/SettingsView.xaml
@@ -1,13 +1,0 @@
-﻿<Page
-    x:Class="WinIRC.SettingsView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:WinIRC"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <ListView x:Name="SettingsList" SelectionMode="None" ItemClick="SettingsList_ItemClick" IsItemClickEnabled="True" />
-    </Grid>
-</Page>
